+++ v2g-liberty/CHANGELOG.md
@@ -1,9 +1,10 @@
 # What's changed?
 
-## 0.8.3 2026-09-??
+## 0.9.0 2026-09-??
 
 ### Fixed
 
+- 🪲 BUG: The app fails to start on unknow new entity (#500)
 - 🪲 BUG: "Battery at max SoC" notification reports the wrong range after a restart (#469)
 - 🪲 BUG: Fix db schema validation (#470)
 - 🪲 BUG: Paused app lets the charger charge the car to full on reconnect (#480, #481)